split off externalptr tests instead of submodule
