+++ tests/extendrtests/man/Wrapper.Rd
@@ -1,21 +1,21 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/extendr-wrappers.R
 \docType{data}
-\name{MySubmoduleClass}
-\alias{MySubmoduleClass}
-\alias{$.MySubmoduleClass}
+\name{Wrapper}
+\alias{Wrapper}
+\alias{$.Wrapper}
 \title{Class for testing (exported)}
 \format{
-An object of class \code{environment} of length 3.
+An object of class \code{environment} of length 11.
 }
 \usage{
-MySubmoduleClass
+Wrapper
 }
 \description{
 Class for testing (exported)
 }
 \examples{
-x <- MySubmoduleClass$new()
+x <- Wrapper$new()
 x$a()
 x$set_a(10)
 x$a()
